+++ py-polars/tests/unit/datatypes/test_schema.py
@@ -667,29 +667,6 @@ def test_alias_prune_in_fold_15438() -> None:
     assert_frame_equal(df, expected)
 
 
-def test_resolved_names_15442() -> None:
-    df = pl.DataFrame(
-        {
-            "x": [206.0],
-            "y": [225.0],
-        }
-    )
-    center = pl.struct(
-        x=pl.col("x"),
-        y=pl.col("y"),
-    )
-
-    left = 0
-    right = 1000
-    in_x = (left < center.struct.field("x")) & (center.struct.field("x") <= right)
-    assert df.lazy().filter(in_x).collect().shape == (1, 2)
-
-
-def test_list_sum_bool_schema() -> None:
-    q = pl.LazyFrame({"x": [[True, True, False]]})
-    assert q.select(pl.col("x").list.sum()).schema["x"] == pl.UInt32
-
-
 @pytest.mark.parametrize("op", ["and_", "or_"])
 def test_bitwise_integral_schema(op: str) -> None:
     df = pl.LazyFrame({"a": [1, 2], "b": [3, 4]})
